+++ src/Lykke.Service.OperationsRepository.Core/CashOperations/ICashOutBase.cs
@@ -23,7 +23,7 @@ namespace Lykke.Service.OperationsRepository.Core.CashOperations
         public string AccountId { get; set; }
         public CashOutVolumeSize VolumeSize { get; set; }
         public string PreviousId { get; set; }
-        public decimal FeeSize { get; set; }
+        public double FeeSize { get; set; }
         public FeeType FeeType { get; set; }
     }
